+++ src/main/java/org/jenkinsci/plugins/cloudhubdeployer/CloudHubRequest.java
@@ -7,8 +7,10 @@ import org.jenkinsci.plugins.cloudhubdeployer.common.RequestMode;
 import org.jenkinsci.plugins.cloudhubdeployer.data.AppInfoJson;
 import lombok.Getter;
 import lombok.Setter;
+import org.jenkinsci.plugins.cloudhubdeployer.data.AutoScalePolicy;
 
 import java.io.PrintStream;
+import java.util.List;
 
 /**
  * CloudHub Request Type
@@ -21,6 +23,9 @@ public class CloudHubRequest {
     @Setter
     @Getter
     private String apiDomainName;
+    @Setter
+    @Getter
+    private String orgId;
     @Setter
     @Getter
     private String envId;
@@ -66,6 +71,12 @@ public class CloudHubRequest {
     @Setter
     @Getter
     private StandardCredentials cloudhubCredentials;
+    @Setter
+    @Getter
+    private boolean autoScalePolicyEnabled;
+    @Setter
+    @Getter
+    private List<AutoScalePolicy> autoScalePolicy;
 
     @Override
     public String toString() {
@@ -92,4 +103,5 @@ public class CloudHubRequest {
         return "{\"username\" : \""+ passwordCredentials.getUsername() +"\" , " +
                 "\"password\" : \""+passwordCredentials.getPassword()+"\" }";
     }
+
 }