+++ src/main/java/org/jenkinsci/plugins/cloudhubdeployer/CloudHubRequestFactory.java
@@ -1,12 +1,19 @@
 package org.jenkinsci.plugins.cloudhubdeployer;
 
 import com.cloudbees.plugins.credentials.common.StandardCredentials;
+import com.google.common.base.Strings;
+import hudson.FilePath;
 import org.jenkinsci.plugins.cloudhubdeployer.common.DebugMode;
 import org.jenkinsci.plugins.cloudhubdeployer.common.RequestMode;
 import org.jenkinsci.plugins.cloudhubdeployer.data.AppInfoJson;
+import org.jenkinsci.plugins.cloudhubdeployer.data.AutoScalePolicy;
+import org.jenkinsci.plugins.cloudhubdeployer.exception.CloudHubRequestException;
 import org.jenkinsci.plugins.cloudhubdeployer.exception.ValidationException;
+import org.jenkinsci.plugins.cloudhubdeployer.utils.DeployHelper;
 
+import java.io.IOException;
 import java.io.PrintStream;
+import java.util.List;
 
 /**
  * Factory to create <code>Cloudhub Request</code> objects
@@ -38,6 +45,11 @@ public class CloudHubRequestFactory {
 
     public CloudHubRequestFactory withCloushubCredentials(final StandardCredentials cloudhubCredentials) {
         cloudHubRequest.setCloudhubCredentials(cloudhubCredentials);
+        return this;
+    }
+
+    public CloudHubRequestFactory withOrgId(String orgId) {
+        cloudHubRequest.setOrgId(orgId);
         return this;
     }
 
@@ -86,11 +98,31 @@ public class CloudHubRequestFactory {
         return this;
     }
 
-    public CloudHubRequestFactory withFilePath(final String filePath) {
-        cloudHubRequest.setFilePath(filePath);
+    public CloudHubRequestFactory withAutoScalepolicy(Boolean enableAutoScalePolicy, List<AutoScalePolicy> autoScalePolicy) {
+        cloudHubRequest.setAutoScalePolicyEnabled(enableAutoScalePolicy);
+        cloudHubRequest.setAutoScalePolicy(autoScalePolicy);
         return this;
     }
 
+    public CloudHubRequestFactory withFilePath(FilePath workspace, final String filePath) {
+        FilePath[] filePaths;
+        try {
+
+            if(null == filePath)
+                return null;
+
+            filePaths = workspace.list(filePath);
+
+            if(filePaths.length > 0){
+                cloudHubRequest.setFilePath(filePaths[0].getRemote());
+            }
+
+        } catch (IOException | InterruptedException e) {
+            e.printStackTrace();
+        }
+
+        return this;
+    }
 
     public CloudHubRequest build() throws ValidationException {
         this.validate();
@@ -99,7 +131,7 @@ public class CloudHubRequestFactory {
 
     private boolean validate() throws ValidationException {
 
-        if (null == this.cloudHubRequest.getApiDomainName()) {
+        if (Strings.isNullOrEmpty(cloudHubRequest.getApiDomainName())) {
             throw new ValidationException("Empty domain name");
         }
 
@@ -107,11 +139,15 @@ public class CloudHubRequestFactory {
             throw new ValidationException("Empty credentials Id");
         }
 
-        if (null == this.cloudHubRequest.getEnvId()) {
+        if (Strings.isNullOrEmpty(cloudHubRequest.getOrgId())) {
+            throw new ValidationException("Empty organization Id");
+        }
+
+        if (Strings.isNullOrEmpty(cloudHubRequest.getEnvId())) {
             throw new ValidationException("Empty environment Id");
         }
 
-        if (null == this.cloudHubRequest.getFilePath()) {
+        if (Strings.isNullOrEmpty(cloudHubRequest.getFilePath())) {
             throw new ValidationException("Empty file path");
         }
 
@@ -139,6 +175,11 @@ public class CloudHubRequestFactory {
             throw new ValidationException("Invalid request mode");
         }
 
+        if(this.cloudHubRequest.isAutoScalePolicyEnabled()){
+            DeployHelper.validateAutoScalePolicy(this.cloudHubRequest.getAutoScalePolicy());
+        }
+
         return true;
     }
+
 }