+++ src/main/java/org/jenkinsci/plugins/cloudhubdeployer/utils/JsonHelper.java
@@ -4,6 +4,7 @@ import com.google.gson.Gson;
 import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
+import org.jenkinsci.plugins.cloudhubdeployer.exception.CloudHubRequestException;
 
 
 public final class JsonHelper {
@@ -13,7 +14,7 @@ public final class JsonHelper {
     }
 
     public static String parseAccessToken(String loginResponse){
-        return new Gson().fromJson(loginResponse, JsonObject.class).get(Constants.LABEL_ACCESS_TOKEN).getAsString();
+        return new Gson().fromJson(loginResponse, JsonObject.class).get(Constants.JSON_KEY_ACCESS_TOKEN).getAsString();
     }
 
     public static boolean checkIfApiExists(String response, String domainName) {
@@ -31,5 +32,31 @@ public final class JsonHelper {
 
     private JsonHelper() {
         // hide constructor
+    }
+
+    public static JsonArray getenvList(String response) throws CloudHubRequestException {
+
+        if(!checkIfKeyExists(response,"data"))
+            throw new CloudHubRequestException("No environment exists with given name or id");
+
+        return new Gson().fromJson(response, JsonObject.class).get("data").getAsJsonArray();
+    }
+
+    public static String verifyOrGetEnvId(String response, String envIdOrName) throws CloudHubRequestException {
+        JsonArray data = getenvList(response);
+
+        for (JsonElement jsonElement : data) {
+            JsonObject jsonObject = jsonElement.getAsJsonObject();
+
+            String id = jsonObject.get("id").getAsString().toLowerCase();
+            String name = jsonObject.get("name").getAsString().toLowerCase();
+
+            if(id.equals(envIdOrName.toLowerCase()) || name.equals(envIdOrName.toLowerCase())){
+                return id;
+            }
+
+        }
+
+        return null;
     }
 }