+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Cloudhub Deployment Plugin Changelog
 
+## Version 0.4.0, 2020-07-21
+* Added support to deploy using environment name
+* Added support for autoscale policies
+* bugfixes and improvement
+
 ## Version 0.1.0, 2020-06-23
 * Initial release
 * Support API deploy, update, restart and delete to Cloudhub